Fixed crash when opening non-existing recent file. Replaced QThread::isRunning() with QThread::wait() where possible. Added QThread::wait() to scripting engine.

+++ src/core/file.h
@@ -19,14 +19,14 @@ public:
 	virtual ~CAFile();
 	
 	inline const int status() { return _status; }
-	inline const float progress() { return _progress; }
+	inline const int progress() { return _progress; }
 	virtual const QString readableStatus() = 0;
 	void setStreamFromFile( const QString filename );
 	void setStreamToFile( const QString filename );
 	
 protected:
 	inline void setStatus( const int status ) { _status = status; }
-	inline void setProgress( const float progress ) { _progress = progress; }
+	inline void setProgress( const int progress ) { _progress = progress; }
 	
 	inline QTextStream *stream() { return _stream; }
 	inline virtual void setStream( QTextStream *stream ) { _stream = stream; }
@@ -35,8 +35,8 @@ protected:
 	inline void setFile( QFile *file ) { _file = file; }
 	
 private:
-	int _status;
-	float _progress;
+	int _status;          // status number
+	int _progress;        // percentage of the work already done
 	QTextStream *_stream;
 	QFile *_file;
 };
